Move baud rate calculation into MIDI_init()

+++ firmware/uart.h
@@ -3,7 +3,7 @@
 
 #include "system.h"
 
-void UART_init(uint8_t);
+void UART_init(uint32_t);
 
 void UART_send(uint8_t);
 uint8_t UART_available(void);
